Enable use of tracer in strict mode
Without this modification one gets the following:

$ node --use_strict app.js
tracer/lib/dailyfile.js:25
            mode: 0666
                  ^^^^

SyntaxError: Octal literals are not allowed in strict mode.

So this commit is not about fixing an error, but being more friendly for
people running Node.js applications in strict mode and in general being
more ready for the future.